Fix integer overflow in av1_caq_select_segment
In av1_caq_select_segment(), sb64_target_rate * xmis * ymis was
evaluated using 32-bit signed integer arithmetic before being cast
to int64_t. When sb64_target_rate is large, this causes signed
integer overflow.

Cast sb64_target_rate to int64_t before multiplying with xmis and ymis,
and declare target_rate as int64_t to prevent overflow when num / denom
exceeds INT_MAX.

Bug: 559079132

+++ av1/encoder/aq_complexity.c
@@ -141,10 +141,10 @@ void av1_caq_select_segment(const AV1_COMP *cpi, MACROBLOCK *mb, BLOCK_SIZE bs,
 
   // Rate depends on fraction of a SB64 in frame (xmis * ymis / bw * bh).
   // It is converted to bits << AV1_PROB_COST_SHIFT units.
-  const int64_t num = (int64_t)(cpi->rc.sb64_target_rate * xmis * ymis)
+  const int64_t num = ((int64_t)cpi->rc.sb64_target_rate * xmis * ymis)
                       << AV1_PROB_COST_SHIFT;
   const int denom = cm->seq_params->mib_size * cm->seq_params->mib_size;
-  const int target_rate = (int)(num / denom);
+  const int64_t target_rate = num / denom;
   double logvar;
   double low_var_thresh;
   const int aq_strength = get_aq_c_strength(cm->quant_params.base_qindex,
